+++ src/lib/goal.ts
@@ -1,21 +1,22 @@
+import { compare } from 'natural-orderby';
 import { Goal } from 'src/lib/api/goal.ts';
 import { GOAL_CADENCE_UNIT_INFO, GOAL_TYPE, GoalHabitParameters, GoalTargetParameters } from 'server/lib/models/goal.ts';
 import { formatCount } from './tally.ts';
 
+const cmp = compare();
 export function cmpGoal(a: Goal, b: Goal) {
   if(a.starred !== b.starred) {
     return a.starred ? -1 : 1;
   }
 
-  if(a.endDate < b.endDate) {
-    return -1;
-  } else if(a.endDate > b.endDate) {
-    return 1;
-  }
+  // TODO: add this back in with a dropdown; as-is, it's not intuitive how this is sorted
+  // if(a.endDate < b.endDate) {
+  //   return -1;
+  // } else if(a.endDate > b.endDate) {
+  //   return 1;
+  // }
 
-  const aTitle = a.title.toLowerCase();
-  const bTitle = b.title.toLowerCase();
-  return aTitle < bTitle ? -1 : aTitle > bTitle ? 1 : 0;
+  return cmp(a.title, b.title);
 }
 
 export function describeGoal(goal: Goal) {