+++ src/lib/tag.ts
@@ -1,0 +1,7 @@
+import { compare } from "natural-orderby";
+import { Tag } from "@prisma/client";
+
+const cmp = compare();
+export function cmpTag(a: Tag, b: Tag) {
+  return cmp(a.name, b.name);
+}