+++ src/stores/board.ts
@@ -1,6 +1,7 @@
 import { defineStore } from 'pinia';
 
 import { getBoards, ExtendedBoard } from 'src/lib/api/board.ts';
+import { cmpBoard } from 'src/lib/board.ts';
 
 export const useBoardStore = defineStore('board', {
   state: () : { boards: ExtendedBoard[] | null; } => {
@@ -19,7 +20,7 @@ export const useBoardStore = defineStore('board', {
     async populate(force = false) {
       if(force || this.boards === null) {
         const boards = await getBoards();
-        this.boards = boards;
+        this.boards = boards.sort(cmpBoard);
       }
     }
   },