+++ src/stores/tag.ts
@@ -1,5 +1,6 @@
 import { defineStore } from 'pinia';
 import { getTags, Tag } from 'src/lib/api/tag.ts';
+import { cmpTag } from 'src/lib/tag';
 
 export const useTagStore = defineStore('tag', {
   state: () : { tags: Tag[] | null; } => {
@@ -9,7 +10,7 @@ export const useTagStore = defineStore('tag', {
     async populate() {
       if(this.tags === null) {
         const tags = await getTags();
-        this.tags = tags;
+        this.tags = tags.sort(cmpTag);
       }
     }
   },