+++ src/stores/work.ts
@@ -1,6 +1,7 @@
 import { defineStore } from 'pinia';
 
 import { getWorks, Work } from 'src/lib/api/work.ts';
+import { cmpWorkByTitle } from 'src/lib/work';
 
 export const useWorkStore = defineStore('work', {
   state: () : { works: Work[] | null; } => {
@@ -19,7 +20,7 @@ export const useWorkStore = defineStore('work', {
     async populate(force = false) {
       if(force || this.works === null) {
         const works = await getWorks();
-        this.works = works;
+        this.works = works.sort(cmpWorkByTitle);
       }
     }
   },